test: split watch mode inspector tests to sequential
